services/grafana_gen/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

pika==1.3.2
python-dotenv==1.1.1
requests==2.32.5
